Let me start with the DVRP paper.
I've been thinking - maybe it is possible to follow up on what I wrote. Maybe GANs are not an evolutionary dead end. In all competitive games, you cannot just assume that the right thing will be done just because the rules have been set. There is always a need for a referee in order to ensure that game flows smoothly.

Maybe cooperation will turn out to just be meta-adversity.