"10:15am. https://arxiv.org/abs/1812.09584 Bayesian Meta-network Architecture Learning
https://www.biorxiv.org/content/early/2018/12/25/503433
How People Initiate Energy Optimization and Converge on Their Optimal Gaits

This was posted on the RL sub recently.

10:35am. 5/17. The stuff in this paper is really complicated and I have only a loose grasp on it. It really is making me wonder whether I could replace all of this with an ensemble and sampling from a Gaussian, similarly what I would do for the actions. I can't really answer that until I understand the stuff in this paper. That will have to be a subject of future work.

10:50am. 7/17. I guess Russel Stuart was right about not having to feel bad about Bayesian nets having to take longer to train.

But I am really wondering how fair this comparison is. If I put in ensemble linking via Gaussian sampling in the hidden layers I wonder what that would get me? Possibly the belief updates much more cheaply.

I could put that in as a constraint for the different RNNs in a ensemble to square their hidden state updates. That would be something.

11:05am. Done with the DVRL paper. It has convinced me. I'll try chasing the ensemble sampling idea in the hidden layers just to see where it goes. If the RNN is relying too much on memory then ensemble sampling will absolutely break that kind of dynamic.

Let me read the other two.

11:35am. The papers were a waste of time. Yeah, my original view that hierarchical ML is a beast was correct. In the DVRL paper, given the claims the vanilla RNN is using memory they sure are not outperforming it by much.

Forget it. The solution will come to me eventually. It is time to open the first crack.

I will not waste time on papers anymore. I will shake of the distractions and put in a real effort into getting the Gridworld game done today. I need to get over my block."