"1:05pm. It is time to start.
So...

1:10pm. I need to come up with a plan.

First let me think back to the poker game. In hindsight it is a mess. I said I would use it as a template, but the issue with it could easily shown by pointing out the whole `internal_representation` thing.

The way I built the whole thing is by first making the game and fitting the RL scaffolding on top of that.

I think that in the future I would be a lot better off by using the formalisms of the field of RL. Meaning having an explicit transition and reward functions as given. I should start with that and then move from there.

I was leaning towards that yesterday, but my previous experience held me back.

So I botched the poker game in the end? Yes.

I did a brilliant job from a certain perspective, but I could have made my job easier.

I'll do it right this time around then.

All the functionality is already there, I just need to shuffle a few things around. I'll use the Bayesian RL book as reference.

1:20pm. I'll leave that for later, after I am done with everything. For poker in particular, I am going to have to do that in order to feed intermediate states as inputs. With 2 player poker it does not matter, but with multiple players I will have to feed observations of other player's betting to the network.

1:25pm. So by that point I will already be diverging from the formalisms.

But the formalisms barely make sense in certain respects. Like beliefs over states. The states might be so highly dimensional that they are not possible to represent. I still do not understand how to generalize sampling states from beliefs over them.

This is so uncomfortable. I've been at this for almost 4 years and I am still at the beginner level. I barely understand anything at all. But I'll figure it out.

And the first step towards that will be to make this crappy Gridworld game.

Let me remove what I did yesterday...

```
//inl n = 50
//inl Position = type Union.int {from=0; near_to=n}
//inl InternalState = type {row=int64; col=int64} \/ .End
//inl State = type Struct.map Position {row=int64; col=int64} \/ .End
//inl Actions = type .Left \/ .Right

//inl serialize = function
//    | .End -> box State .End
//    | x -> Struct.map Position x |> box State

//inl transition action state =
//    | .Start -> {row=0; col=0}, {reward=0.01 / to float64 n}
//    inl row = row - 1
//    inl col =
//        match action with
//        | .Left ->
//            inl col = col - 1
//            if col < 0 then 0 else col
//        | .Right ->
//            inl col = col + 1
//            if col >= n then n-1 else col

//inl game player =
//    ()
```

I will start anew this time with the reward function. It is fine if the states themselves remain implicit. I can't possibly put every single thing into the RL formalism. I'll do it in the most realistic way possible.

I need to distinguish between real state, believed states and observations."