"1:50pm. Let me make the reward function. Had to take a break.
```
inl reward {row col} =
    if row = col then 0.01 / to float64 n + (if row = n - 1 then 1.0 else 0.0)
    else 0.0
```

Here is the reward of state.

Now comes the transition.

1:55pm.

```
inl n = 50
inl reward {row col} =
    if row = col then 0.01 / to float64 n + (if row = n - 1 then 1.0 else 0.0)
    else 0.0

inl transition action {row col} ret =
    inl row = row + 1
    if row = n then ret .End
    else
        inl col =
            match action with
            | .Left ->
                inl col = col - 1
                if col < 0 then 0 else col
            | .Right ->
                inl col = col + 1
                if col >= n then n-1 else col
        ret {row col}
```

Yeah, this is the old me. This is nice easy. Now let me make the game loop.

2:15pm.

```
inl game player =
    Loops.for' {from=0; near_to=n; state={row=0; col=0}
        body=inl {state i next} ->
            player.observe (observe state)
            player.reward (reward state)
            transition player.action state <| function
            | .End -> player.game_over
            | state -> next state
        finally=ignore
        }
```

In the poker game I modified the state directly in the player. From a certain perspective that makes sense as it simplifies the implementation.

From another it complicates it as the game player and the RL player are not the same thing and it conflates the two. I haven't made a mistake in using mutable state last time. What I did make a mistake though is to not separate the RL and game parts.

That `internal_representation` functions should have rightfully been `observe`.

Let me implement `observe`."