"4pm. This is hardly a short break. But I am just so blocked.
I am extremely dissatisfied with the current state of the libraries.

Right now I have absolutely no idea how implement that piece of crap tabular player.

And I know deep down that every little design change will take hours and hours of working through type errors. I really wish I was working in F# or Ocaml or Haskell or literally any language with a powerful and yet automatic type system. I do not want to think about boxing or any of that other crap. I do not want to write my own type inference.

The problem I have is an easy one. `learning1` and `learning2` instead of just loading data and then feeding it to the network do a lot of handling state and all that other crap.

I made a serious mistake in not making a proper interface for those tests first. I thought it might be a good idea, but decided that finishing the tests takes precedence. This is the same choice I had for the poker game and I did the same thing.

As a result everything I have for RL is now coupled with the poker game. Fuck.

And this is Spiral and not F#, meaning a refactor would break everything and necessitate a lengthy testing period. I am so tired of that crap.

It is ironic, but after all my effort and development I now think that I am a good enough programmer that I do not need Spiral. I'd rather move fast and break things in a language like Haskell. After mastering Spiral intensional polymorphism that language will be a piece of cake. Nonetheless, unless I want to go Python + TF/PT, I will have to keep using for the sake of doing Cuda programming. If Python was a functional PL rather than a toy for brainlets I would honestly consider it.

Hence the ideal move here is not to quit Spiral in favor of something else, but to double and triple up on OO programming.

4:10pm. I actually have a good thing going with what I did with the poker players in the poker game.

I need to double and triple up on the whole OO thing. Make sure that I can just make calls to the interfaces, never expose the details of that to the user.

In Spiral that goes that much more than in any other language as it forces the user to do custom type inference. If any language needs OO, it is Spiral.

4:15pm. I am not sure I feel like programming anymore.

Let me move all the poker stuff to the old module. It is disgusting and just having it there is throwing off my focus. Every two minutes I think of how I will fit the new thing together with it.

4:20pm. It is kind of ironic all things considered.

This exploration thing should be trivial to implement and yet it will take me two weeks most likely.

It is not the algorithm that is the challenge, but the programming.

The Gridworld itself is not a challenge.

The true challenge is making the interfaces. This is one thing that I've overlooked up to now and hence failed.

I've learned so many, many thing in programming.

I guess it is only right that here at the end of this old journey, I will finally learn the value of mutable data structures.

4:25pm. It is not Spiral that is the problem - it cannot possibly be. It certainly has the power of abstraction. Rather, what is shackling me down are my own bad choice whether that be library design or research direction. I've wasted so much time this year implementing irrelevant bullshit.

Yet, I should not mistake my position. Sure, I might be in the pit of hell right now, but I am not at my weakest at the moment. I am at my strongest.

4:30pm. New goal: Forget everything else. Forget even about the Gridworld. Solving it is not the issue.

Rather redo the tabular player from scratch. This time do it perfectly. Make it the correct template for the NN agents.

Then after that I will redo the `Learning` tests. I will create the correct interfaces so that if I ever need another test that uses NNs, it will be trivial to make.

Let me step away from the screen for a while. I just need time to gather my determination.

Languor is an enemy to be pierced by the belief in a brighter future. Once I carve it out, I will be able to use it. I will be able to show my true power."