"10:05am. Where was I? I finally got up. Well, let me chill for a bit. I am sort of having a break.
I should be programming, but I do not feel like it. What I will aim for is to completely rebuild the tabular player, but failing that, I'll just step away from the screen to brainstorm.

Programming is like a wave, the desire to do it has a beginning and an end. I need to build up the wave in myself. Just forget the past. Forget everything that I did. Bring out all the power and this time do it properly.

10:15am. Iron Ladies is fun even though the author has the worst sense of scale that I've ever seen. Right now I am just chilling.

It is kind of like that - even though I've written all the code myself, the poker game and the tests are just dead weight to me right now. I'd rather just forget they exist so I can go on with my (programming) life, and I that is what I am angling for.

I am pleased with the state of the ML library though. The rough parts that still remain like the PG function are not a big deal. I will have to make `seq_iter`, but that should be an interesting task.

Even though I made a lot of failed experiments that I regret, during that time I did take the opportunity to evolve Spiral and the ML library. After an extra year of work Spiral itself is in good shape despite it slowness. Yes, it is unsuitable for experimentation - I really stabbed myself on that, and refactoring the library code really makes me want to shot myself, but it is the best language in the world to cultivate a program.

Can I believe that? Yes, that has not changed. The motivation for making the language and the benefits it has over others has not changed. It is true that it enables an unique style of programming no other language in existence can.

10:30am. Can I still get better at programming?"