```             inl action = player.act {state=observe state; reward=reward state}             transition action state <| function             | .End -> player.game_over             | state -> next state ```
I am just thinking about this right now.

```
            player.observe (observe state)
            player.reward (reward state)
            transition player.action state <| function
            | .End -> player.game_over
            | state -> next state
```

It is really just quite difficult to give a reward for just being in a state. I've always thought of reward being also dependent on the action rather than just the state.

This whole conception is messing with me.

I think I am going to go with my own formulation rather than the nonsensical problem formulation which gives the actor a negative reward in the initial state.