"11:05am.
```
inl game player =
    Loops.for' {from=0; near_to=n; state={row=0; col=0}
        body=inl {state i next} ->
            inl observation = observe state
            inl action = player.act observation
            transition action state <| function
            | .End -> player.game_over
            | state ->
                inl reward = reward state
                player.remember {observation action reward}
                next state
        finally=ignore
        }
```

Yeah, this is fine. I am going to change `State` to `Observation` inside the players. Saying that something is `State` is completely nonsensical to me. Forget that I am using `State` everywhere. I will do it properly from now on.

11:10am. Now how about the template for the player?

```
//inl player_tabular_sarsa {Action State init learning_rate} =
//    inl {action} = PlayerTabular {Action State init learning_rate}
//    inl Reward = type float64
//    inl History = type State \/ Action \/ Reward
//    inl buffer = ResizeArray.create {elem_type=History}
//    inl trace = ResizeArray.create {elem_type=heap (action State .bck)}

//    inl methods = {
//        observe=inl s rep -> s.data.buffer.add (box History rep)
//        reward=inl s v -> s.data.buffer.add (box History v)
//        action=inl s ->
//            inl hist = get_latest_slice s.data.buffer
//            inl {action bck} = s.data.action rep
//            s.data.trace.add (heap bck)
//            action
//        game_over=inl s ->
//            s.data.trace.foldr (inl bck v -> bck v) (dyn (to float32 v)) |> ignore; s.data.trace.clear
//        }

//    Object
//        .member_add methods
//        .data_add {name; win=ref 0; action trace buffer}
```

Let me put this trash here for reference. I will do it from scratch. One thing I am going to do is add eligibility traces based on what I have currently in the ML library. It is a really good formulation, I will give it that.

11:30am. I am getting distracted, but somehow I am starting to get that familiar feeling of being in the zone.

```
inl player_tabular_sarsa {Observation Action} =
    inl Observation = Observation \/ ()
    inl Action = Action \/ ()
    inl elem_type = type {observation=Observation; action=Action; reward=float64}

    inl trace = ResizeArray.create {elem_type}
```

I've been thinking about the lines of adding unit to the types, but no. This would be wrong. Rather, what I should do is remove `Action`.

Rather than do sarsa why not also do value iteration?

I could mix and match that way. I want this so I can pass observations of what other players are doing to the player even when it is not its turn to act. This will be needed for multiplayer games.

Yeah I see it now. That way I could propagate value from observations without actions to observations with actions.

What is amazing about this is that yesterday when I thought about this, my plan was just to ignore the observations without actions entirely. But now I can clearly see the value in propagating the information through them.

This is what I am going to aim for then. It is going to take a complete overhaul of how agents work though. But that is what I am here for. To do it right this time.

11:35am. Throw away the past and take control. That is an important part of programming.

It is important to understand that even though they might be millions of lines of code in a codebase, the foundations that lie below it is much smaller. And all the action is truly in mastering those foundations.

Let me have breakfast here. I'll resume later."