"1:25pm. Let me see if I can get into the zone again. A little break and I am completely out of it.
...And I see my first issue. If I allow all kinds of stuff as observation and action I will have a problem, namely that I won't know how to search over actions.

There is also an issue of using {observation action} as input to Union. Namely the action will come last and it would really be better that it came first for more efficient memory accessing.

```
    | _ :: _ -> Tuple.foldl prod (0,1) x
    | .(_) | () -> 0,1
    | {!block} -> module_foldl (const prod) (0,1) x
```

No wait, that was wrong. So action should come first. Damn, that is complicated. I messed up the Union. The innermost dimension should always be to the right.

...

```
inl from_dense true_is ty ar =
    inl fatal_fail conv _ = failwith conv "The array is in the wrong format."

    inl rec from_dense ty i =
        inl peek i {on_succ on_fail} = if true_is (ar i) then on_succ i else on_fail ()
        inl find {from near_to} {on_succ on_fail} =
            Loops.for' {from near_to
                body=inl {next i} -> peek i {on_succ on_fail=next}
                finally=on_fail
                }

        inl {s ty=conv} = ty
        match ty with
        | {union_type=x} ->
            inl next _ = function
                | .find -> Option.none conv
                | .found, v -> Option.some v

            Tuple.foldr (inl x next i state ->
                inl r = indiv join stack (from_dense x i)
                match state with
                | .find ->
                    match r with
                    | .Some, v -> next (i + x.s) (.found, box conv v)
                    | .None -> next (i + x.s) state
                | .found, v ->
                    match r with
                    | .Some, v -> fatal_fail (Option.none conv) ()
                    | .None -> next (i + x.s) state
                ) x next i .find
        | {x} ->
            match x with
            | _ :: _ ->
                Tuple.foldr (inl x next i l on_fail ->
                    inl r = indiv join stack (from_dense x i)
                    match r with
                    | .Some, v -> next (i + x.s) (v :: l) (fatal_fail (Option.none conv))
                    | .None -> on_fail()
                    ) x (inl i l on_fail -> Option.some (Tuple.rev l)) i () (inl _ -> Option.none conv)
            | .(_) | () -> peek i {on_succ=(inl _ -> Option.some x); on_fail=(inl _ -> Option.none x)}
            | {!block} ->
                module_foldr (inl k x next i m on_fail ->
                    inl r = indiv join stack (from_dense x i)
                    match r with
                    | .Some, v -> next (i + x.s) (module_add k v m) (fatal_fail (Option.none conv))
                    | .None -> on_fail()
                    ) x (inl i m on_fail -> Option.some m) i {} (inl _ -> Option.none conv)
            | {from=.(from) near_to=.(near_to) value block=()} ->
                find {from=i; near_to=i+near_to-from} {
                    on_succ=inl i' ->
                        find {from=i'+1; near_to} {
                            on_succ=fatal_fail (Option.none conv)
                            on_fail=inl _ -> Option.some ({x with value=i' - i + from})
                            }
                    on_fail=inl _ -> Option.none conv
                    }

    match from_dense ty 0 with
    | .Some, v -> v
    | .None -> fatal_fail ty.ty ()
```

Yeah, I am not going to be redesigning all of this, that is for sure. I've just added a note for the time being.

1:40pm. Agh, I should have thought about this when I was first making it, but I did not and now it is too late to fix it. The code looks like somebody else wrote it at this point. I know that it was not trivial to write this.

But at any rate, it is convenient as I can write `{observation action}` and have it be in the correct order.

1:45pm. There is another issue that I need to take care of as well.

Originally I thought that I would make the dictionary a one big union type...like `{observation=Position; action=Orientation} \/ {observation=Other; action=SomethingElse}`. I thought I would use an array for this whole big thing, but now that I am going deeper I see an issue with this.

Namely that I won't be able to return the action due to them having different types.

There is an easy fix for this - split the types into individual ones, use a dictionary for the observation (rather than a array) and use an array for the action. This is what I've been using for the existing tabular player, except this time I'll have more dictionaries.

It is fine, this will be a piece of cake.

https://www.newyorker.com/science/elements/how-the-artificial-intelligence-program-alphazero-mastered-its-games

Let me take a break here. I want to watch the video by Kasparov anyway.

2:10pm. Let me do the chores. I had some time to think about it. Yeah, breaks and progress are going to be outsized until I catch my wind.

2:45pm. Done with chores. I wonder how it took me so long?

At any rate, let me resume. Let me see if I can catch wind again."