"5:30pm.
```
        body=inl {state i next} ->
            inl action = player.act (observe state)
            transition action state <| function
            | .End -> player.optimize()
            | state ->
                player.reward (reward state)
                next state
```

This is quite clean now.

5:40pm. I am just thinking about what is next now.

Game specific interfaces would be one.

Well, let me just enjoy this moment for a bit and then I'll call it a day.

I am rather please with how things turned out. I would not describe the kind of programming I did today as intense. It was rather tranquil actually. This kind of programming is perfectly acceptable. Piece by piece, I will get there. All I need to do is one more push and I will have succeeded in my goal of solving the Sea game with a tabular player.

5:45pm. As the last thing for the day, let me put the player in its own module.

6pm. Done. I put in some final touches.

Unlike with the library refactoring, the testing for this won't be too bad. 15-30m at most.

After I am done with the tabular player on the Sea game, I will move on to redoing the two learning tests.

It is time to eat. I am done for the day."