"12:45pm. Done with breakfast. It is time to start.
Agh, I do not feel like it. I really do not feel like it, but I must. Let me get rid of the distractions for a few hours and focus solely on programming.

The talk by Martha White was interesting even if there is no firm conclussion why distributional losses work better apart from improving the gradient norm. This really sets up the stage for what I am going to introduce.

The HRL talk by the sweaty Indian dude is just filler. I skimmed it since it was just describing what HRL is. What was really looking for are hints on how to do it and the lecture was not offering me anything on that front.

Well, it was not a complete waste of time, but now it is time to get serious.

12:55pm. My focus is narrowing. And I feel like doing it as my mind homes in on the task at hand. Great.

Yesterday I did the template. And I could put that player in and it would do the right thing, but I am missing a few pieces at the moment. Namely I need something to record the average rewards for the player and print them out later. I need to make sure that the thing is actually solving the game. Also I need to print out the sequence as things go along. I need to make sure that it is going right.

1pm. Well, let me give it a try. Printing stuff out won't be too much of a problem. For logging rewards I'll just use an exponential moving average. Actually, I do not even need that since once it learns the optimal policy it will always get 0.99.

1:15pm.

```
inl player =
    inl {act reward optimize} = PlayerTabular.template {init=2f32; elem_type={Observation Action}; learning_rate=0.01f32; trace=0.5f32; discount=1f32}
    inl act {obs with row col} =
        inl row, col = row.value, col.value
        Console.write {row col}
        act obs
    inl reward_sum = ref 0.0
    inl reward rew =
        reward_sum := reward_sum() + rew
        reward rew
    inl optimize x =
        Console.writeline ""
        Console.printfn "The sum of rewards for the episode is {0}." (reward_sum())
        reward_sum := 0f32
        optimize x
    heap {act reward optimize}
```

This should do it.

How about I just put this in a loop and get it to work. I've reduced the n to 10. Later I'll add a counter to print it only ever 100 iterations or such.

1:25pm.

```
inl player =
    inl {act reward optimize} = PlayerTabular.template {init=2f32; elem_type={Observation Action}; learning_rate=0.01f32; trace=0.5f32; discount=1f32}
    inl act {obs with row col} =
        inl row, col = row.value, col.value
        Console.write {row col}
        act obs
    inl reward_sum = ref 0.0
    inl reward rew =
        reward_sum := reward_sum() + rew
        reward rew
    inl optimize x =
        Console.writeline ""
        Console.printfn "The sum of rewards for the episode is {0}." (reward_sum())
        reward_sum := 0f32
        optimize x
    heap {act reward optimize}

inl num_episondes = 100
Loops.for {from=0; near_to=num_episodes; body=inl {i} -> game player}
```

This should do it. Now comes the debugging gauntlet. I'll go at it one by one. Let me take care of this and then come the learning tests. Once I've clean up everything thoroughly I'll bring in the AC player. I'll clean it up and then upgrade it with the ensembling idea.

At that point I will be ready challenge the future.

Even with deep exploration the Singularity will still be far away, but at least I will be able to kick some pebbles away from my path.

It is time to take care of this."