"``` The sum of rewards for the episode is 0.001111. [0, 0][1, 0][2, 0][3, 1][4, 0][5, 1][6, 0][7, 1][8, 2][9, 1] The sum of rewards for the episode is 0. [0, 0][1, 1][2, 2][3, 3][4, 4][5, 5][6, 4][7, 3][8, 2][9, 3] The sum of rewards for the episode is 0.005555. [0, 0][1, 0][2, 1][3, 2][4, 3][5, 4][6, 3][7, 2][8, 3][9, 4] The sum of rewards for the episode is 0. [0, 0][1, 1][2, 0][3, 0][4, 1][5, 2][6, 1][7, 2][8, 1][9, 0] The sum of rewards for the episode is 0.001111. [0, 0][1, 0][2, 0][3, 1][4, 2][5, 3][6, 2][7, 3][8, 4][9, 3] The sum of rewards for the episode is 0. ```
Oh, whops. Moving along the diagonal should give negative rewards.

```
The sum of rewards for the episode is 0.
[0, 0][1, 1][2, 2][3, 3][4, 4][5, 5][6, 4][7, 3][8, 4][9, 5]
The sum of rewards for the episode is -0.005555.
[0, 0][1, 0][2, 0][3, 0][4, 1][5, 2][6, 1][7, 2][8, 1][9, 0]
The sum of rewards for the episode is 0.
[0, 0][1, 1][2, 0][3, 1][4, 2][5, 3][6, 2][7, 3][8, 2][9, 1]
The sum of rewards for the episode is -0.001111.
```

Yeah, it is fine now. Let me stop printing out individual episodes and only do the print every 100 episodes.

1:55pm.

```
[0, 0][1, 0][2, 1][3, 2][4, 3][5, 4][6, 5][7, 6][8, 7][9, 8]
The sum of rewards for the episode is 0.
```

Hmmm, are 10k episodes not enough to deal with this? Let me increase the LR.

```
[0, 0][1, 0][2, 0][3, 0][4, 1][5, 0][6, 0][7, 0][8, 0][9, 0]
The sum of rewards for the episode is 0.
```

This is after 1M steps. Something is wrong. Let me set n=3. Now it is n=10.

Even with n=3 it fails. Let me print out every individual step.

```
The sum of rewards for the episode is 0.99.
[0, 0][1, 0][2, 1]
The sum of rewards for the episode is 0.
[0, 0][1, 1][2, 2]
The sum of rewards for the episode is 0.99.
[0, 0][1, 0][2, 0]
The sum of rewards for the episode is 0.
[0, 0][1, 0][2, 1]
The sum of rewards for the episode is 0.
[0, 0][1, 1][2, 0]
The sum of rewards for the episode is -0.005.
```

So it does hit the rightmost one. Just what is happening to the values then.

2:05pm.

```
inl {act reward optimize} = PlayerTabular.template {init=1.2f32; elem_type={Observation Action}; learning_rate=0.01f32; trace=0.5f32; discount=1f32}
```

Let me set the init to something more sensible. Let me also print it out properly.

2:10pm.

```
The sum of rewards for the episode is 0.99.
{V = 0.2403323; a = 1; obs = [2, 2]}
{V = 1.19558; a = 1; obs = [1, 1]}
{V = 0.07475109; a = 1; obs = [0, 0]}

The sum of rewards for the episode is 0.
{V = 0.2195477; a = 0; obs = [2, 0]}
{V = 0.2206566; a = 0; obs = [1, 0]}
{V = 0.07472417; a = 0; obs = [0, 0]}

The sum of rewards for the episode is 0.
{V = 0.2195477; a = 0; obs = [2, 1]}
{V = 0.2204537; a = 1; obs = [1, 0]}
{V = 0.07453134; a = 0; obs = [0, 0]}
```

There is something wrong with this. The values cannot possibly be propagated correctly. Should I disable the trace?

Let me run it for 10k episodes."