"2pm. Done with the chores. It is time for me to finally start programming. I caught up with Kengan Asura and now I have nothing more to distract me. This was a long 3h break. From now I will just focus on programming until the end of the day.
Yeah, I know. My effort has been mediocre so far, but I am not the same as I was when I first started this and had seemingly limitless motivation and energy. Once I am done with testing the new exploration method, and if it turns out to work then that will be a massive boost to motivation. At that point I will be able to put in sore Real Effort.

Right now, I think that it should work and there are not reason why it should not, but it is not like I can prove it ahead of time. I can't really reason through an optimization process like I could a regular program. I put a lot of faith into differentiable plasticity and Zap and Rudder, and various other things and they all turned out to be flops. Even those inverse covariance matrix updates were not handled correctly. I spent the last 8 months essentially burning myself up. I moved and trusted that the wind would lift me up, but it never did.

2:10pm. Apart from making deep exploration work, hierarchical RL is something that will have to be done competently, but I do not have any clue as to how to do unsupervised learning properly. The GVFNs are a powerful indication, but there are not the only one. There are all sorts of indication pointing in various directions and I can't tell the dominant perspective.

It is going to be hard. There are hints scattered around already, but they are scattered too much for me. I am dumb so I am going to need some more handholding in order to figure it out, much like with anchored ensembling.

I should have figured out the deep exploration method I have now as soon as I saw Gaussian processes, but I did not.

3 years are a lot of time to turn things around. I will study and keep my eyes out for any hints and then put it all together.

The field is moving quickly. Without a doubt, the winds of the future will blow the scattered pieces until my vision grasps them.

2:20pm. The future is the future. In the present, let me do the design for `learning2`."