"5:20pm. I do not feel like resuming right now, but I think I still have another hour or two left in me.
I do not have to test it today. How about I actually do the agents for the two learning tests? The RL one won't be hard after that.

Let me sketch out the solution for the two learning agents. I'll save bolting down the details and dealing with the mutable cost thing for tomorrow.

5:30pm.

```
inl agent =
    Agent.feedforward
        .with_context s
        .initialize network (train_images 0)
        .with_rate rate
        .with_error Error.softmax_cross_entropy
```

Yeah, I really do not feel like doing this, but I am going to keep going regardless for the sake of my desire until the ideas stop coming and my inspiration dries up.

5:40pm.

```
    inl Agent =
        inl feedforward =
            inl methods = {
                with_context = inl s cd -> s.data_add {cd}
                with_rate = inl s rate -> s.data_add {rate}
                with_error = inl s rate -> s.member_add {error}
                initialize = inl s network sample ->
                    inl cd = s.data.cd
                    inl side = sample.span_outer
                    s.data_add {network = init cd network size}
                }

        inl recurrent =
            inl methods = {
                with_context = inl s cd -> s.data_add {cd}
                with_rate = inl s rate -> s.data_add {rate}
                with_error = inl s rate -> s.member_add {error}
                initialize = inl s network sample ->
                    inl cd = s.data.cd
                    inl side = sample.span_outer
                    inl s = s.data_add {network = init cd network size}

                    // Here comes the inference and the ResizeArray.
                }
        {feedforward recurrent rl}
```

I really want to quit here. There is no way I feel like implementing the `dyn`ing of the offset or the mutable cost function here.

But let me at least put in the two infer functions.

5:50pm. https://arxiv.org/abs/1704.00109
Snapshot Ensembles - Train 1, get M for free

I'll save this paper for later.

6pm.

```
    inl Agent =
        inl feedforward =
            inl methods = {
                with_context = inl s cd -> s.data_add {cd}
                with_rate = inl s rate -> s.data_add {rate}
                with_error = inl s rate -> s.member_add {error}
                initialize = inl s network sample ->
                    inl cd = s.data.cd
                    inl side = sample.span_outer
                    s.data_add {network = init cd network size}
                }

        inl recurrent =
            inl methods = {
                with_context = inl s cd -> s.data_add {cd}
                with_rate = inl s rate -> s.data_add {rate}
                with_error = inl s rate -> s.member_add {error}
                initialize = inl s network sample ->
                    inl cd = s.data.cd
                    inl side = sample.span_outer
                    inl network = init cd network size
                    inl s = s.data_add {network}

                    inl elem_type, {run truncate} =
                        Union.infer {
                            run={
                                map=inl {network cd} {input label} ->
                                    inl network, input = run cd input network
                                    inl {out bck} = final label input cd
                                    inl out _ = cd.CudaTensor.get out |> to float64 // TODO: Do the mutable cost thing here.
                                    inl prev = {out bck=Struct.map (inl {bck} -> bck) network, bck}
                                    inl network = Struct.map (inl x -> {x without bck}) network
                                    {network cd prev}
                                input=const {input=sample; label=sample} // TODO: `dyn` the offset for sample.
                                block=()
                                }
                            truncate={
                                map=inl {network cd} _ ->
                                    inl cd = cd'.RegionMem.create
                                    inl network =
                                        Struct.map (function
                                            | {state} as d ->
                                                inl state =
                                                    Struct.map
                                                        (inl out -> out.update_body (inl {x with ar} -> cd.RegionMem.assign ar.ptr; x))
                                                        {(indiv state) without weights}
                                                    |> heap
                                                {d without bck with state}
                                            | d -> {d without bck}
                                            ) network
                                    cd'.RegionMem.clear
                                    cd.refresh
                                    {network cd}
                                input=const ()
                                block=()
                                }
                            } {network cd}

                    inl ar = ResizeArray.create {elem_type}
                    ()
                }
        {feedforward recurrent}
```

I am getting way too ahead of myself with the recurrent stuff. Damn, this code is just nasty in all sorts of ways, but I cannot do without it.

Just so I do not have to look at it, I am going to take the recurrent stuff out of the library for now. Let it be backed up in this journal until I am done with `learning1`.

6:05pm. I just do not feel like doing it anymore. Getting `learning1` done will require me to deal with the output stuff and to implement all the additional functions. I cannot help not wanting to put any more work for the day.

I am both at my lowest and my highest right now. The jolt I have given myself in 2014 has been overdone, and I need some positive reinforcement to get the blood flowing. I just picked the wrong path in the last 8 months and am now paying the price. Had I started with the anchored ensembling idea and stuck to dealing with exploration I'd definitely be motivated to put in effort.

But well, it is fine. If you cannot put in 8, put in 5. If you cannot put in 5, then put in 3 and so on. The most important thing is to put in some time into it each and every day.

I think that doing it 8-10 hours per day like I did in the early years might have been overdoing it, but I should be able to move from the current 3 to 5-6 at least.

6:10pm. It is fine. I will sacrifice a few days just to do this redesign properly. I'll do the tests and this time, they will stay solved for good. Then I will do the actor critic and it will stay solved for good after that.

Then I will take a few days to implement ensembling and deep exploration.

In a week I will know exactly where my future prospects lie. If I can grasp this then I will be able to grasp anything.

Now it is time for lazing."